Original change: https://googleplex-android-review.googlesource.com/c/platform/packages/modules/AdServices/+/25096783

+++ adservices/tests/unittest/service-core/src/com/android/adservices/service/signals/SignalsEncodingE2ETest.java
@@ -307,7 +307,10 @@ public class SignalsEncodingE2ETest {
 
     @After
     public void teardown() {
-        mEncoderPersistenceDao.deleteAllEncoders();
+        if (mEncoderPersistenceDao != null) {
+            mEncoderPersistenceDao.deleteAllEncoders();
+        }
+
         if (mStaticMockSession != null) {
             mStaticMockSession.finishMocking();
         }
